requirements.txt
annotated-types>=0.7.0
anyio>=4.9.0
certifi>=2025.4.26
click>=8.1.8
fastapi>=0.115.12
GitPython>=3.1.40
h11>=0.16.0
httpcore>=1.0.9
httpx>=0.28.1
idna>=3.10
packaging>=25.0
psutil>=7.0.0
pydantic>=2.11.5
pydantic_core==2.33.2
PyQt6>=6.9.0
PyQt6-Qt6>=6.9.0
PyQt6_sip>=13.10.2
python-dotenv>=1.1.0
sniffio>=1.3.1
starlette>=0.46.2
typing-inspection>=0.4.1
typing_extensions>=4.13.2
uvicorn>=0.34.2
python-rtmidi>=1.5.8
